+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-0.18.0 (unreleased)
+0.18.0 (2015-10-04)
 *******************
 
 Features:
@@ -10,6 +10,8 @@ Features:
 * Add ``fields.DelimitedList`` (:issue:`66`). Thanks :user:`jmcarp`.
 * ``TornadoParser`` will parse json with ``simplejson`` if it is installed.
 * ``BottleParser`` caches parsed json per-request for improved performance.
+
+No breaking changes. Yay!
 
 0.17.0 (2015-09-29)
 *******************